Separate toggle state and status in GitListEntry
